Move BuildMetadataQuery
